requirements.txt
openai
openai-agents
flask
twilio
flask-sockets
gevent
gevent-websocket
python-dotenv
numpy
pyaudio